+++ src/main/java/org/zalando/nakadi/repository/kafka/KafkaLocationManager.java
@@ -9,9 +9,10 @@ import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.util.StringUtils;
 import org.zalando.nakadi.repository.zookeeper.ZooKeeperHolder;
 
-import java.io.UnsupportedEncodingException;
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Properties;
@@ -27,7 +28,13 @@ public class KafkaLocationManager {
 
     public KafkaLocationManager(final ZooKeeperHolder zkFactory, final KafkaSettings kafkaSettings) {
         this.zkFactory = zkFactory;
-        this.kafkaProperties = buildKafkaProperties(fetchBrokers());
+        if (StringUtils.isEmpty(kafkaSettings.getBootstrapServers())) {
+            this.kafkaProperties = buildKafkaProperties(fetchBrokers());
+        } else {
+            this.kafkaProperties = new Properties();
+            this.kafkaProperties.put(CommonClientConfigs.BOOTSTRAP_SERVERS_CONFIG,
+                    kafkaSettings.getBootstrapServers());
+        }
         this.kafkaSettings = kafkaSettings;
     }
 
@@ -40,8 +47,8 @@ public class KafkaLocationManager {
             this.port = port;
         }
 
-        static Broker fromByteJson(final byte[] data) throws JSONException, UnsupportedEncodingException {
-            final JSONObject json = new JSONObject(new String(data, "UTF-8"));
+        static Broker fromByteJson(final byte[] data) throws JSONException {
+            final JSONObject json = new JSONObject(new String(data, StandardCharsets.UTF_8));
             final String host = json.getString("host");
             final Integer port = json.getInt("port");
             return new Broker(host, port);
@@ -73,19 +80,19 @@ public class KafkaLocationManager {
 
     private static String buildBootstrapServers(final List<Broker> brokers) {
         final StringBuilder builder = new StringBuilder();
-        brokers.stream().forEach(broker -> builder.append(broker).append(","));
+        brokers.forEach(broker -> builder.append(broker).append(","));
         return builder.deleteCharAt(builder.length() - 1).toString();
     }
 
-    private Properties buildKafkaProperties(final List<Broker> brokers) {
+    private static Properties buildKafkaProperties(final List<Broker> brokers) {
         final Properties props = new Properties();
-        props.put("bootstrap.servers", buildBootstrapServers(brokers));
+        props.put(CommonClientConfigs.BOOTSTRAP_SERVERS_CONFIG, buildBootstrapServers(brokers));
         return props;
     }
 
     @Scheduled(fixedDelay = 30000)
     private void updateBrokers() {
-        if (kafkaProperties != null) {
+        if (StringUtils.isEmpty(kafkaSettings.getBootstrapServers()) && kafkaProperties != null) {
             final List<Broker> brokers = fetchBrokers();
             if (!brokers.isEmpty()) {
                 kafkaProperties.setProperty(CommonClientConfigs.BOOTSTRAP_SERVERS_CONFIG,